IBL and PBR model loading fixes

+++ RXNEngine/src/RXNEngine/Renderer/Framebuffer.h
@@ -44,7 +44,7 @@ namespace RXNEngine {
 
 	struct FramebufferSpecification
 	{
-		uint32_t Width = 0, Height = 0;
+		float Width = 0, Height = 0;
 		FramebufferAttachmentSpecification Attachments;
 		uint32_t Samples = 1;
